fix: Use hooks instead of deface
