--- app/overrides/account/login-3.4.rb
@@ -1,5 +1,0 @@
-Deface::Override.new virtual_path: 'account/login',
-                     name: 'add-maintenance-message-to-login-3.4',
-                     original: '31fdd230fc8fa8802fad742b3f2436c5c2a4e017',
-                     insert_top: '#login-form',
-                     partial: 'maintenance/maintenance_login_3_4'
