--- app/overrides/layouts/base.rb
@@ -1,5 +1,0 @@
-Deface::Override.new :virtual_path  => 'layouts/base',
-                     :name          => 'add-maintenance-message-to-bodytop',
-                     :original      => 'b85b7e064a792d811d7060c206f55f08797aef91',
-                     :insert_top    => 'body',
-                     :partial       => 'maintenance/maintenance_body'
