--- app/views/maintenance/_maintenance_login_3_3.html.erb
@@ -1,11 +1,0 @@
-<%
-  settings = MaintenanceModeFunctions.get_maintenance_plugin_settings
-%>
-<% if settings[:maintenance_active] || MaintenanceModeFunctions.is_now_scheduled_maintenance %>
-<tr class="maintenance_message">
-  <td colspan="2">
-    <%= textilizable(settings[:maintenance_message_f]) %>
-    <p><%= l(:maintenance_mode_login_admin_only) %></p>
-  </td>
-</tr>
-<% end %>
