--- app/views/maintenance/_maintenance_login_3_4.html.erb
@@ -1,9 +1,0 @@
-<%
-  settings = MaintenanceModeFunctions.get_maintenance_plugin_settings
-%>
-<% if settings[:maintenance_active] || MaintenanceModeFunctions.is_now_scheduled_maintenance %>
-<div class="maintenance_message">
-  <%= textilizable(settings[:maintenance_message_f]) %>
-  <p><%= l(:maintenance_mode_login_admin_only) %></p>
-</div>
-<% end %>
