+++ app/views/settings/_redmine_maintenance_mode_settings.html.erb
@@ -22,7 +22,7 @@
 
 <hr/>
 
-<div id="settings_maintenance_schedule"<%= raw MaintenanceModeFunctions.is_now_scheduled_maintenance ? ' class="active"' : '' %>>
+<div id="settings_maintenance_schedule"<%= raw MaintenanceMode::Functions.is_now_scheduled_maintenance ? ' class="active"' : '' %>>
 	<br/>
 	
 	<%= raw l(:maintenance_mode_schedule_directions) %>
