+++ lib/maintenance_mode.rb
@@ -8,11 +8,11 @@ module MaintenanceMode
       # show maintenance message for all "normal" users (except admins)
       def show_maintenance_mode_page
         # read plugin settings
-        settings = MaintenanceModeFunctions.get_maintenance_plugin_settings
+        settings = MaintenanceMode::Functions.get_maintenance_plugin_settings
 
         # only activate maintenance message if maintenance mode is activated or
         # if we're in the middle of a scheduled maintenance
-        if (settings[:maintenance_active] || MaintenanceModeFunctions.is_now_scheduled_maintenance) &&
+        if (settings[:maintenance_active] || MaintenanceMode::Functions.is_now_scheduled_maintenance) &&
            !(User.current.admin? || (Redmine::Plugin.installed?('redmine_sudo') && User.current.sudoer?))
           logout_user if User.current.logged?
           require_login unless params[:controller] == 'account' && params[:action] == 'login'
